Modifed useRoutes
Modified useRoutes to get routes.tsx which hold the data about links, even if the hook it's still useful sometime it can be good like for our "home page" to grab the link directly so we keep SSR in some page/component.
