--- src/app/sharedTypes.ts
@@ -1,8 +1,0 @@
-export interface LinkData {
-	text: string;
-	href: string;
-	style?: React.CSSProperties;
-}
-
-export type LinksData = Array<LinkData>
-
